Move language server files to 'activation/languageServer' folder
