+++ news/3 Code Health/2756.md
@@ -1,0 +1,1 @@
+Moved languageServer-related files to a languageServer folder.
