+++ src/client/activation/languageServer/languageServerFolderService.ts
@@ -6,13 +6,13 @@
 import { inject, injectable } from 'inversify';
 import * as path from 'path';
 import * as semver from 'semver';
-import { EXTENSION_ROOT_DIR } from '../common/constants';
-import { log } from '../common/logger';
-import { NugetPackage } from '../common/nuget/types';
-import { IFileSystem } from '../common/platform/types';
-import { IConfigurationService } from '../common/types';
-import { IServiceContainer } from '../ioc/types';
-import { FolderVersionPair, IDownloadChannelRule, ILanguageServerFolderService, ILanguageServerPackageService } from './types';
+import { EXTENSION_ROOT_DIR } from '../../common/constants';
+import { log } from '../../common/logger';
+import { NugetPackage } from '../../common/nuget/types';
+import { IFileSystem } from '../../common/platform/types';
+import { IConfigurationService } from '../../common/types';
+import { IServiceContainer } from '../../ioc/types';
+import { FolderVersionPair, IDownloadChannelRule, ILanguageServerFolderService, ILanguageServerPackageService } from '../types';
 
 const languageServerFolder = 'languageServer';
