+++ src/client/activation/languageServer/languageServerPackageRepository.ts
@@ -4,8 +4,8 @@
 'use strict';
 
 import { inject, injectable } from 'inversify';
-import { AzureBlobStoreNugetRepository } from '../common/nuget/azureBlobStoreNugetRepository';
-import { IServiceContainer } from '../ioc/types';
+import { AzureBlobStoreNugetRepository } from '../../common/nuget/azureBlobStoreNugetRepository';
+import { IServiceContainer } from '../../ioc/types';
 
 const azureBlobStorageAccount = 'https://pvsc.blob.core.windows.net';
